pyproject.toml
[project]
name = "quickbooks-mcp"
version = "0.1.0"
description = "MCP server for QuickBooks Online — invoices, customers, vendors, bills, payments, accounts, reports, and more"
requires-python = ">=3.10"
dependencies = [
    "dedalus-mcp",
    "pydantic",
    "python-dotenv",
]

[project.optional-dependencies]
dev = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
